+++ dragonfly/player/bossbar/bossbar.go
@@ -9,7 +9,7 @@ import (
 // it. The health shown by the bar may be changed.
 type BossBar struct {
 	text   string
-	health float32
+	health float64
 }
 
 // New creates a new boss bar with the text passed. The text is formatted according to the rules of
@@ -27,7 +27,7 @@ func (bar BossBar) Text() string {
 // WithHealthPercentage sets the health percentage of the boss bar. The value passed must be between 0 and 1.
 // If a value out of that range is passed, SetHealthPercentage panics.
 // The new BossBar with the changed health percentage is returned.
-func (bar BossBar) WithHealthPercentage(v float32) BossBar {
+func (bar BossBar) WithHealthPercentage(v float64) BossBar {
 	if v < 0 || v > 1 {
 		panic("boss bar: value out of range: health percentage must be between 0.0 and 1.0")
 	}
@@ -37,7 +37,7 @@ func (bar BossBar) WithHealthPercentage(v float32) BossBar {
 
 // HealthPercentage returns the health percentage of the boss bar. The number returned is a value between 0
 // and 1, with 0 being an empty boss bar and 1 being a full one.
-func (bar BossBar) HealthPercentage() float32 {
+func (bar BossBar) HealthPercentage() float64 {
 	return bar.health
 }
